+++ tests/test_db_setup.py
@@ -1,8 +1,4 @@
-import pytest
-
-from django.conf import settings
-
-from .db_helpers import mark_exists, mark_database, drop_database, db_exists
+from .db_helpers import mark_exists, mark_database, drop_database, db_exists, skip_if_sqlite
 
 
 def test_db_reuse(django_testdir):
@@ -11,9 +7,7 @@ def test_db_reuse(django_testdir):
     to be available and the environment variables PG_HOST, PG_DB, PG_USER to
     be defined.
     """
-
-    if settings.DATABASES['default']['ENGINE'] == 'django.db.backends.sqlite3':
-        pytest.skip('Do not test db reuse since database does not support it')
+    skip_if_sqlite()
 
     django_testdir.create_test_module('''
 import pytest
@@ -57,3 +51,51 @@ def test_db_can_be_accessed():
 
     # Make sure the database has been re-created and the mark is gone
     assert not mark_exists()
+
+
+def test_xdist_with_reuse(django_testdir):
+    skip_if_sqlite()
+
+    drop_database('gw0')
+    drop_database('gw1')
+
+    django_testdir.create_test_module('''
+import pytest
+
+from .app.models import Item
+
+def _check(settings):
+    # Make sure that the database name looks correct
+    db_name = settings.DATABASES['default']['NAME']
+    assert db_name.endswith('_gw0') or db_name.endswith('_gw1')
+
+    assert Item.objects.count() == 0
+    Item.objects.create(name='foo')
+    assert Item.objects.count() == 1
+
+
+@pytest.mark.django_db
+def test_a(settings):
+    _check(settings)
+
+
+@pytest.mark.django_db
+def test_b(settings):
+    _check(settings)
+
+''')
+
+    result = django_testdir.runpytest('-vv', '-n2', '-s', '--reuse-db')
+    result.stdout.fnmatch_lines(['*PASSED*test_a*'])
+    result.stdout.fnmatch_lines(['*PASSED*test_b*'])
+
+    assert db_exists('gw0')
+    assert db_exists('gw1')
+
+    result = django_testdir.runpytest('-vv', '-n2', '-s', '--reuse-db')
+    result.stdout.fnmatch_lines(['*PASSED*test_a*'])
+    result.stdout.fnmatch_lines(['*PASSED*test_b*'])
+
+    result = django_testdir.runpytest('-vv', '-n2', '-s', '--reuse-db', '--create-db')
+    result.stdout.fnmatch_lines(['*PASSED*test_a*'])
+    result.stdout.fnmatch_lines(['*PASSED*test_b*'])